GS-8525 fix to scale by number of cpu cores client side implementation failures.
svn path=/trunk/openspaces/; revision=80832


Former-commit-id: 125e5303defa99fe33a8e891de26cad7771f97a8

+++ src/main/src/org/openspaces/admin/pu/elastic/ElasticMachineProvisioningConfig.java
@@ -3,7 +3,7 @@ package org.openspaces.admin.pu.elastic;
 import org.openspaces.admin.bean.BeanConfig;
 
 public interface ElasticMachineProvisioningConfig extends BeanConfig {
-
-    int getMinimumNumberOfCpuCoresPerMachine();
+    
+    double getMinimumNumberOfCpuCoresPerMachine();
 
 }
